refactor:(app) separate sections from components
