+++ app/src/sections/Booking/FormSection/CancellationPolicyModal.jsx
