+++ app/src/sections/Booking/FormSection/RulesModal.jsx
